move cache health check to diff file
